refactor: LottoStatistics 내부에서 총 당첨 금액 합 계산

+++ src/main/java/lotto/global/constant/ConsoleType.java
@@ -26,7 +26,7 @@ public enum ConsoleType {
         return comment;
     }
 
-    public String getComment(Integer input) {
+    public String getComment(Long input) {
         return String.format(comment, input);
     }
